Add support for --enable-static-build

+++ src/hotspot/share/compiler/disassembler.cpp
@@ -72,7 +72,7 @@ bool Disassembler::load_library() {
   char* p = strrchr(buf, '/');
   *p = '\0';
   strcat(p, "/lib/");
-  lib_offset = jvm_offset = strlen(buf);
+  lib_offset = jvm_offset = (int)strlen(buf);
 #else
   {
     // Match "jvm[^/]*" in jvm_path.
